requirements.txt
# Core ML & Deep Learning
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# Computer Vision
opencv-python>=4.6.0
Pillow>=9.0.0
scikit-image>=0.19.0

# Data Science
numpy>=1.21.0
pandas>=1.4.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0

# API & Web
fastapi>=0.78.0
uvicorn[standard]>=0.18.0
python-multipart>=0.0.5

# Jupyter
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
tqdm>=4.64.0
PyYAML>=6.0
requests>=2.28.0

# Additional ML libraries
tensorboard>=2.9.0
wandb>=0.12.0
albumentations>=1.2.0
timm>=0.6.0
efficientnet-pytorch>=0.7.1

# Model export
onnx>=1.12.0
onnxruntime>=1.12.0

# Explainability
grad-cam>=1.4.0

# Monitoring
mlflow>=1.28.0
prometheus-client>=0.14.0

# Development
black>=22.0.0
flake8>=4.0.0
pytest>=7.0.0
pre-commit>=2.19.0
